refactoring to avoid circular import issues
