+++ learning_resources/management/commands/backpopulate_micromasters_data.py
@@ -3,9 +3,9 @@
 from django.core.management import BaseCommand
 
 from learning_resources.etl.constants import ETLSource
+from learning_resources.etl.utils import resource_delete_actions
 from learning_resources.models import LearningResource
 from learning_resources.tasks import get_micromasters_data
-from learning_resources.utils import resource_delete_actions
 from main.utils import now_in_utc
